pyproject.toml
[project]
name = "ml-bird-classification"
version = "0.1.0"
description = "Media Metadata Enrichment Pipeline — bird species classification from live camera streams"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "Pillow>=10.2",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "tensorboard>=2.16",
    "boto3>=1.34",
    "confluent-kafka>=2.3",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "elasticsearch>=8.12",
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "ffmpeg-python>=0.2",
    "opencv-python-headless>=4.9",
    "prometheus-client>=0.20",
    "imagehash>=4.3",
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "apscheduler>=3.10",
    "asyncpg>=0.29",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
spark = [
    "pyspark>=3.5",
]
flink = [
    "apache-flink>=1.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
